refactor: extract request-dispatch internals into app_serve.go
